Faltu commit

+++ layup/src/layers.hpp
@@ -61,8 +61,8 @@ public:
 
     /* Layup Addon */
     LayupLayer_t layer_type;
-    float input_size = 0.0f;
-    float flops = 1.0f;
+    double input_size = 0.0;
+    double flops = 1.0;
 
 protected:
     /** Previous layer. */
